requirements.txt
# Core packages
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<0.17.0
transformers>=4.30.0,<5.0.0
datasets>=2.12.0,<3.0.0
numpy>=1.23.0,<2.0.0
pillow>=9.4.0,<10.0.0
matplotlib>=3.7.0,<4.0.0
scikit-learn>=1.2.0,<2.0.0

# Additional dependencies
argparse>=1.4.0  # Python standart kütüphanesiyle birlikte gelir, ancak eski sistemler için
tqdm>=4.65.0,<5.0.0  # Genellikle transformers/datasets ile kullanılır

accelerate>=0.20.0,<1.0.0  # Dağıtılmış eğitim için